requirements.txt
jupyterlab
notebook
mlflow
pandas
numpy
matplotlib
scikit-learn
statsmodels
scipy
cloudpickle
pickle
treamlit
joblib